chore,refac: utils: Re-organize utilities
- Add: `._utils` submodule.
- Change: Move definitions from `.utils` to `._utils`
- Change: Export only public utilities from `.utils`.
- Change: Update affected modules and tests.

  All modules now import from `._utils` instead of `.utils`.
