+++ src/term_image/__init__.py
@@ -26,9 +26,9 @@ from operator import truediv
 
 from typing_extensions import ClassVar, Final
 
-from . import utils
+from . import _utils
+from ._utils import arg_value_error_range, get_cell_size
 from .exceptions import TermImageError
-from .utils import arg_value_error_range, get_cell_size
 
 version_info = (0, 8, 0, "dev")
 
@@ -37,7 +37,7 @@ __version__ = ".".join(map(str, version_info[:3]))
 if version_info[3:]:
     __version__ += "-" + ".".join(map(str, version_info[3:]))
 
-DEFAULT_QUERY_TIMEOUT: Final[float] = utils._query_timeout
+DEFAULT_QUERY_TIMEOUT: Final[float] = _utils._query_timeout
 """Default timeout for :ref:`terminal-queries`
 
 .. seealso:: :py:func:`set_query_timeout`.
@@ -90,7 +90,7 @@ def disable_queries() -> None:
     NOTE:
         This affects all :ref:`dependent features <queried-features>`.
     """
-    utils._queries_enabled = False
+    _utils._queries_enabled = False
 
 
 def disable_win_size_swap() -> None:
@@ -104,10 +104,10 @@ def disable_win_size_swap() -> None:
         This affects :ref:`auto-cell-ratio` computation and size computations for
         :ref:`graphics-based`.
     """
-    if utils._swap_win_size:
-        utils._swap_win_size = False
-        with utils._cell_size_lock:
-            utils._cell_size_cache[:] = (0,) * 4
+    if _utils._swap_win_size:
+        _utils._swap_win_size = False
+        with _utils._cell_size_lock:
+            _utils._cell_size_cache[:] = (0,) * 4
 
 
 def enable_queries() -> None:
@@ -118,12 +118,12 @@ def enable_queries() -> None:
     NOTE:
         This affects all :ref:`dependent features <queried-features>`.
     """
-    if not utils._queries_enabled:
-        utils._queries_enabled = True
-        getattr(utils.get_fg_bg_colors, "_invalidate_cache")()
-        getattr(utils.get_terminal_name_version, "_invalidate_cache")()
-        with utils._cell_size_lock:
-            utils._cell_size_cache[:] = (0,) * 4
+    if not _utils._queries_enabled:
+        _utils._queries_enabled = True
+        getattr(_utils.get_fg_bg_colors, "_invalidate_cache")()
+        getattr(_utils.get_terminal_name_version, "_invalidate_cache")()
+        with _utils._cell_size_lock:
+            _utils._cell_size_cache[:] = (0,) * 4
 
 
 def enable_win_size_swap() -> None:
@@ -137,10 +137,10 @@ def enable_win_size_swap() -> None:
         This affects :ref:`auto-cell-ratio` computation and size computations for
         :ref:`graphics-based`.
     """
-    if not utils._swap_win_size:
-        utils._swap_win_size = True
-        with utils._cell_size_lock:
-            utils._cell_size_cache[:] = (0,) * 4
+    if not _utils._swap_win_size:
+        _utils._swap_win_size = True
+        with _utils._cell_size_lock:
+            _utils._cell_size_cache[:] = (0,) * 4
 
 
 def get_cell_ratio() -> float:
@@ -217,7 +217,7 @@ def set_query_timeout(timeout: float) -> None:
     if timeout <= 0.0:
         raise arg_value_error_range("timeout", timeout)
 
-    utils._query_timeout = timeout
+    _utils._query_timeout = timeout
 
 
 _cell_ratio: float | None = 0.5
