+++ src/term_image/widget/_urwid.py
@@ -12,9 +12,9 @@ from .. import _ctlseqs as ctlseqs
 
 # These sequences are used during performance-critical operations that occur often
 from .._ctlseqs import BEGIN_SYNCED_UPDATE, END_SYNCED_UPDATE, ESC_b, SGR_NORMAL_b
+from .._utils import arg_type_error, get_terminal_name_version, lock_tty, write_tty
 from ..exceptions import UrwidImageError
 from ..image import BaseImage, ITerm2Image, KittyImage, Size, TextImage
-from ..utils import arg_type_error, get_terminal_name_version, lock_tty, write_tty
 
 # NOTE: Any new "private" attribute of any subclass of an urwid class should be
 # prepended with "_ti" to prevent clashes with names used by urwid itself.
