+++ tests/__init__.py
@@ -62,17 +62,17 @@ def toggle_is_on_kitty():
     is_on_kitty = not is_on_kitty
 
 
-term_image.utils.get_terminal_size = get_terminal_size
+term_image._utils.get_terminal_size = get_terminal_size
 
 terminal_name_version = ("", "")
-term_image.utils.get_terminal_name_version = get_terminal_name_version
+term_image._utils.get_terminal_name_version = get_terminal_name_version
 
 cell_size = None
 term_image.get_cell_size = get_cell_size
-term_image.utils.get_cell_size = get_cell_size
+term_image._utils.get_cell_size = get_cell_size
 
 fg_bg = [(0, 0, 0), (0, 0, 0)]
-term_image.utils.get_fg_bg_colors = get_fg_bg_colors
+term_image._utils.get_fg_bg_colors = get_fg_bg_colors
 
 import term_image.image  # noqa: E402
